Do not use SANE environment to set errno in one-argument math functions.
Many <math.h> functions previously examined the exception flags to determine whether to set errno. The problem with this is that SANE calls may set exception flags, but they never clear them (except via explicit calls to do that). So a SANE exception flag may be set based on any previous SANE call, not just the last one. This could result in errno being set erroneously when is should not be. In a few calls like sinh and cosh, this kind of issue could also affect the computational logic, so incorrect results could be produced.

This patch fixes the issue for all the one-argument <math.h> functions in SysFloat. The new logic sets errno to EDOM if the output is NAN and the input was not, or to ERANGE if the output is +-INF and the input was not. This handles the possible domain errors, pole errors, and overflows for most math functions. It does not flag underflows, but that is not required and was not generally done by the old code either.

This also fixes sqrt(negative) to set errno to EDOM rather than ERANGE.

Some of the multi-argument <math.h> functions have the same kind of issues. They will be addressed later.

This fixes #3 and also fixes #4.