Fix fmod to return correct values when second argument is negative.
It would generally give a value with the wrong sign in these cases, e.g. for fmod(11.5, -3.0) or fmod(-11.5, -3.0).

This also changes the behavior if the second argument is zero: it is now treated as a domain error, which is one of the options specified in the standard for that case.