Simplify modf and do not set errno in it.
The errno-setting code used the SANE environment, creating the same problem as in other functions. I just removed it, because the standards to not require modf to set errno, and the only way to get an overflow would be to pass an input value that is outside the range of double.